Revert "feat: refactor SignUp logics to signUpWiewModel"
This reverts commit 1a3dc7229096eb30ceeba7713e91c8bbdca49a9e.
